Move balance handling code into MarketMaker
